+++ pasqal_cloud/job.py
@@ -1,8 +1,10 @@
 from typing import Any, Dict, List, Optional
 
 from pydantic import BaseModel, Extra
+from requests import HTTPError
 
 from pasqal_cloud.client import Client
+from pasqal_cloud.errors import JobCancellingError
 
 
 class Job(BaseModel):
@@ -21,7 +23,10 @@ class Job(BaseModel):
         - errors: Error messages that occurred while processing job.
         - start_timestamp: The timestamp of when the job began processing.
         - end_timestamp: The timestamp of when the job finished processing.
-        - result: Result of the job.
+        - full_result: Dictionnary of all the results obtained after complete execution
+            of the job. It maps the type of results (e.g. "counter", "raw")
+            to the associated execution result.
+        - result: Bitstring counter result. Should be equal to `full_results["counter"]`
         - variables: Dictionary of variables of the job.
             None if the associated batch is non-parametrized.
         - group_id (deprecated): Use project_id instead.
@@ -50,6 +55,9 @@ class Job(BaseModel):
 
     def cancel(self) -> Dict[str, Any]:
         """Cancel the current job on the PCS."""
-        job_rsp = self._client._cancel_job(self.id)
+        try:
+            job_rsp = self._client._cancel_job(self.id)
+        except HTTPError as e:
+            raise JobCancellingError from e
         self.status = job_rsp.get("status", "CANCELED")
         return job_rsp